- added refresh_token endpoint. - removed auth.py - updated Config to work different platforms e.g.: linux, mac, win. - moved auth to user.py.
