+++ app/config.py
@@ -1,5 +1,6 @@
 import secrets
 from pydantic import BaseSettings
+import os
 
 class Config(BaseSettings):
     db_username: str
@@ -8,8 +9,14 @@ class Config(BaseSettings):
     secret_key: str = secrets.token_hex(32)
 
     class Config:
-        # TODO: make it platform independent
-        env_file = "..\.env"
+         # TODO: make it platform independent
+        from sys import platform
+        if platform == "win32":
+            # Windows...
+            env_file = "..\.env"
+        else:
+            # Linux, mac...
+            env_file = "../.env"
         env_file_encoding = "utf-8"
