+++ app/main.py
@@ -4,7 +4,6 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from api.users import user_routes
 from api.items import item_routes
-from api.auth import auth_routes
 
 origins = [
            "http://localhost:8000",
@@ -14,7 +13,6 @@ origins = [
 app = FastAPI()
 app.include_router(user_routes)
 app.include_router(item_routes)
-app.include_router(auth_routes)
 
 app.add_middleware(
     CORSMiddleware,
